backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
scapy==2.5.0
openai==0.28.1
reportlab==4.0.4
pandas==2.0.3
python-dotenv==1.0.0